vectrofyPython/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
Pillow==9.5.0
opencv-python==4.8.0.74
numpy==1.26.4
gunicorn
